+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + UI and gameplay improvements.
 + Several bug fixes.
 + Code improvements.
++ Added RuleScript documentation
 
 ## 0.11.0
 + Anonymous, basic tracking added to get insightful data about how players play the game.